feat: delete DOProvider.Apply v1 wrapper; release v1.4.0 (Phase 3)
DOProvider.Apply was dead code post-Phase-2.5 v2 cutover (DO declares
ComputePlanVersion=v2; wfctl routes via wfctlhelpers.ApplyPlanWithHooks
+ IaCProviderFinalizer.FinalizeApply, bypassing this method).

Method body replaced with sentinel-error stub returning
ErrApplyV1Removed for diagnostic classification. interfaces.IaCProvider
contract preserved per ADR 0024 — clean deletion not viable without
breaking all 4 plugins simultaneously; interface segregation deferred
to separate refactor design (filed as followup tracking issue).

deferredUpdater interface relocated from provider.go to iacserver.go
(sole consumer is FinalizeApply); preserves Go-level type clarity.

provider_deferred_test.go deleted — regression coverage migrated to
iacserver_finalize_test.go per Phase 2.5 T8 quality-review verdict.

Eliminates FinalizeApply-vs-Apply deferred-flush loop duplication
Copilot flagged on PR #123 (workflow#695 Phase 2.5).

Scope extension (team-lead authorized): delete 11 pre-existing v1 Apply
tests across provider_apply_test.go + provider_test.go that exercise
the now-deleted v1 wrapper dispatch behavior. Equivalent coverage
preserved at wfctlhelpers.ApplyPlan tests + internal/drivers/*_test.go
per-driver tests + iacserver_finalize_test.go (Phase 2.5 T8). Helper
types relocated to internal/deferred_test_helpers_test.go for use by
remaining tests.

Closes workflow#695 followup item 'Phase 3: delete DOProvider.Apply
v1 wrapper'.
